Extracted router in own file.
